+++ sportstracker/src/main/kotlin/de/saring/sportstracker/storage/db/ExerciseRepository.kt
@@ -5,7 +5,7 @@ import de.saring.sportstracker.core.STExceptionID
 import de.saring.sportstracker.data.Exercise
 import de.saring.sportstracker.data.SportType
 import de.saring.sportstracker.storage.db.RepositoryUtil.getEquipmentById
-import de.saring.sportstracker.storage.db.RepositoryUtil.getIntegerOrNull
+import de.saring.sportstracker.storage.db.RepositoryUtil.getLongOrNull
 import de.saring.sportstracker.storage.db.RepositoryUtil.getSportSubTypeById
 import de.saring.sportstracker.storage.db.RepositoryUtil.getSportTypeById
 import de.saring.util.Date310Utils.dateToLocalDateTime
@@ -37,11 +37,11 @@ class ExerciseRepository(
                 while (rs.next()) {
                     val exercise = readFromResultSet(rs)
 
-                    val sportType = getSportTypeById(sportTypes, rs.getInt("SPORT_TYPE_ID"))
+                    val sportType = getSportTypeById(sportTypes, rs.getLong("SPORT_TYPE_ID"))
                     exercise.sportType = sportType
-                    exercise.sportSubType = getSportSubTypeById(sportType, rs.getInt("SPORT_SUBTYPE_ID"))
+                    exercise.sportSubType = getSportSubTypeById(sportType, rs.getLong("SPORT_SUBTYPE_ID"))
 
-                    val equipmentID = getIntegerOrNull(rs, "EQUIPMENT_ID")
+                    val equipmentID = getLongOrNull(rs, "EQUIPMENT_ID")
                     exercise.equipment = if (equipmentID == null) null else getEquipmentById(sportType, equipmentID)
                     exercises.add(exercise)
                 }
@@ -56,7 +56,7 @@ class ExerciseRepository(
         throw UnsupportedOperationException("Use readAll(List<SportType>) for reading all Exercises!")
     }
 
-    override fun readById(entryId: Int): Exercise {
+    override fun readById(entryId: Long): Exercise {
         throw UnsupportedOperationException("Use readById(Int, List<SportType>) for reading an Exercise!")
     }
 
@@ -67,7 +67,7 @@ class ExerciseRepository(
     override val logger: Logger = Logger.getLogger(ExerciseRepository::class.java.name)
 
     override fun readFromResultSet(rs: ResultSet): Exercise {
-        val exercise = Exercise(rs.getInt("ID"))
+        val exercise = Exercise(rs.getLong("ID"))
         exercise.dateTime = dateToLocalDateTime(rs.getDate("DATE_TIME"))
         exercise.intensity = Exercise.IntensityType.valueOf(rs.getString("INTENSITY"))
         exercise.duration = rs.getInt("DURATION")
